refactor: reduce duplication code - create class action remove node and go action
