+++ cigen/core/github/go_action.py
@@ -2,17 +2,14 @@ from __future__ import annotations
 
 from abc import ABC, abstractmethod
 
-import yaml
-
-from cigen.core.github.base_action import base_action, base_version_list_action
-from cigen.core.github.github_action import Steps, OnEvent
+from cigen.core.github.github_action import Steps, Action
 
 
 class GoActionBuilder(ABC):
 
     @property
     @abstractmethod
-    def build(self) -> GoAction:
+    def build(self) -> Action:
         pass
 
     @property
@@ -116,7 +113,7 @@ class GoActionBuilderImpl(GoActionBuilder):
         self.reset_steps()
 
     def reset(self):
-        self._build = GoAction(self.name, self.version, self.on, self.step, self.env)
+        self._build = Action(self.name, self.version, self.on, self.step, self.env)
 
     def reset_steps(self):
         self._steps = GoActionSteps(self.version)
@@ -194,40 +191,10 @@ class GoActionBuilderImpl(GoActionBuilder):
         self._steps.version = param
 
 
-class GoAction:
-    on: OnEvent
-    steps: Steps
-
-    def __init__(self, name, version, on, steps: Steps, env=None) -> None:
-        self.name = name
-        self.version = version
-        self.on = on
-        self.steps = steps
-        self.env = env
-
-    def base(self):
-        return base_action(self.name, self.on, self.steps)
-
-    def base_version_list(self):
-        return base_version_list_action(self.name, self.on, self.steps, self.version)
-
-    def base_to_yaml(self):
-        return yaml.dump(self.base())
-
-    def run(self):
-        return self.base()
-
-    def run_with_env(self):
-        return {
-            **self.base(),
-            'env': self.env
-        }
-
-
 class GoActionSteps:
     def __init__(self, version) -> None:
         if version is None:
-            raise Exception('Version is required')
+            raise TypeError('Version is required')
         self.version = version
 
     @staticmethod
@@ -371,4 +338,3 @@ class ActionCIGenGolang:
 
     def _action_steps_run_build(self):
         self.builder.add_steps(self.builder.step_run_build())
-
